add docstrings finishes[#162050564] finishes[#162050543] finishes[#162050285] finishes[#162050178]

+++ sendapi/routes/user.py
@@ -14,6 +14,10 @@ from validations import Validator
 @app.route("/api/v1/users", methods=['GET'])
 @jwt_required
 def fetch_all_users():
+    """ 
+        Function fetch all users
+        :return success message:
+    """
     current_user=get_jwt_identity() 
     user=User() 
     find_user_with_id=(user.get_user_by_id(current_user))
@@ -27,6 +31,11 @@ def fetch_all_users():
 @app.route("/api/v1/users/<int:userId>/parcels", methods=['GET'])
 @jwt_required
 def fetch_all_parcels_by_user(userId):
+    """ 
+        Function fetch all parcels by user
+        :param userId:
+        :return success message:
+    """
     user_id=get_jwt_identity()
     if not user_id:
         return jsonify({"message":"You are not logged in"}),401
@@ -41,10 +50,17 @@ def fetch_all_parcels_by_user(userId):
 @app.route("/api/v1/users/<int:userId>", methods=['DELETE'])
 @jwt_required
 def delete_user(userId):
+    """ 
+        Function delete user
+        :param userId:
+        :return success message:
+    """
     user_id=get_jwt_identity()
     print(user_id)
     user=User() 
     get_user_by_id=(user.get_user_by_id(user_id))
+    if not get_user_by_id:
+        return jsonify({"message":"You are not a system user"}),401
     print(get_user_by_id)
     if get_user_by_id['role'] =='admin':
         user.delete_user(userId)
@@ -55,11 +71,18 @@ def delete_user(userId):
 @app.route("/api/v1/users/<int:userId>", methods=['GET'])
 @jwt_required
 def get_user(userId):
+    """ 
+        Function get user
+        :param userId:
+        :return user details:
+    """
     user_id=get_jwt_identity()
     if user_id!=userId:
         return jsonify({"message":"You can only view your details"})
     user=User() 
     get_user_by_id=(user.get_user_by_id(user_id))
+    if not get_user_by_id:
+        return jsonify({"message":"You are not a registered user"}),401
     if get_user_by_id['role'] =='admin' and user_id==userId:
         return jsonify(get_user_by_id),200
     return jsonify({"message":"only admin can access user details"}),401
